feat: pseudo boolean encoding of feature attribute constraints that contain doubles by encoding them in integer values

+++ src/main/java/de/vill/model/expression/NumberExpression.java
@@ -87,7 +87,7 @@ public class NumberExpression extends Expression {
     @Override
     public List<Literal> getAsSum(List<PBConstraint> additionalConstraints) {
         List<Literal> result = new LinkedList<>();
-        result.add(new Literal((int) number, null));
+        result.add(new Literal(number, null));
         return result;
     }
